requirements.txt
selenium
webdriver-manager
plyer
msedge-selenium-tools